\name{BGFD-package}
\alias{BGFD-package}
\docType{package}
\title{
Fitting of Bell-G and Complementary Bell-G family of Distributions
}

\description{
Evaluates the probability density function, cumulative distribution function, quantile function, random numbers,
survival function, hazard rate function, and maximum likelihood estimates for the following distributions:
Bell exponential, Bell extended exponential, Bell Weibull, Bell extended Weibull, Bell-Fisk, Bell-Lomax, Bell Burr-XII,
Bell Burr-X, complementary Bell exponential, complementary Bell extended exponential, complementary Bell Weibull,
complementary Bell extended Weibull, complementary Bell-Fisk, complementary Bell-Lomax, complementary Bell Burr-XII
and complementary Bell Burr-X distribution.
}

\details{
\tabular{ll}{
Package: \tab BGFD\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2023-04-06\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}, Muhammad Imran <imranshakoor84@yahoo.com> and M.H. Tahir <mht@iub.edu.pk>.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Fayomi, A., Tahir, M. H., Algarni, A., Imran, M. and Jamal, F. (2022). A new useful exponential model with applications to quality control and
actuarial data. Computational Intelligence and Neuroscience, 2022.

Alanzi, A. R., Imran, M., Tahir, M. H., Chesneau, C., Jamal, F., Shakoor, S. and Sami, W. (2023). Simulation analysis,
properties and applications on a new Burr XII model based on the Bell-X functionalities, AIMS Mathematics, 8(3): 6970-7004.

Algarni, A. (2022). Group Acceptance Sampling Plan Based on New Compounded Three-Parameter Weibull Model.
Axioms, 11(9): 438.
}

\note{
Acknowledgments:

Muhammad Imran would like to express his acknowledgments to Dr. Farukh Jamal, his PhD co-supervisor,
and the Islamia University of Bahawalpur, Pakistan.
}

\keyword{Parameter induction}
\keyword{Compounding}
\keyword{Distribution theory}
\keyword{Generalization of distribution }


